modified:   Linux Bash Scripts/Scripts/Anaconda/Anaconda.sh modified:   Linux Bash Scripts/Scripts/AutoStart/AutoStart.sh modified:   Linux Bash Scripts/Scripts/Bitwarden/closeopen.py modified:   Linux Bash Scripts/Scripts/Clementine/Volume/text.txt modified:   Linux Bash Scripts/Scripts/Clementine/Volume/text2.txt modified:   Linux Bash Scripts/Scripts/Clementine/Volume/text3.txt modified:   Linux Bash Scripts/Scripts/Clementine/check.txt modified:   Linux Bash Scripts/Scripts/Clementine/vis.txt modified:   Linux Bash Scripts/Scripts/Clementine/visible.txt modified:   Linux Bash Scripts/Scripts/DesktopChange/text.txt modified:   Linux Bash Scripts/Scripts/Discord/check.txt modified:   Linux Bash Scripts/Scripts/Discord/vis.txt modified:   Linux Bash Scripts/Scripts/Discord/visible.txt modified:   Linux Bash Scripts/Scripts/DuplicatedEntries/text.txt modified:   Linux Bash Scripts/Scripts/DuplicatedEntries/text3.txt modified:   Linux Bash Scripts/Scripts/GPU/text.csv modified:   Linux Bash Scripts/Scripts/GPU/text.txt modified:   Linux Bash Scripts/Scripts/Jupyter/Jupyter.sh modified:   Linux Bash Scripts/Scripts/Mail/check.txt modified:   Linux Bash Scripts/Scripts/Mail/vis.txt modified:   Linux Bash Scripts/Scripts/Mail/visible.txt modified:   Linux Bash Scripts/Scripts/Memory/text.txt modified:   Linux Bash Scripts/Scripts/Memory/text2.txt modified:   Linux Bash Scripts/Scripts/Memory/text3.txt modified:   Linux Bash Scripts/Scripts/Messenger/check.txt modified:   Linux Bash Scripts/Scripts/Messenger/vis.txt modified:   Linux Bash Scripts/Scripts/Messenger/visible.txt modified:   Linux Bash Scripts/Scripts/MicVolume/Mic.py modified:   Linux Bash Scripts/Scripts/MicVolume/MicVolume.sh modified:   Linux Bash Scripts/Scripts/Noise/Noise_c.sh modified:   Linux Bash Scripts/Scripts/Notion/check.txt modified:   Linux Bash Scripts/Scripts/Notion/vis.txt modified:   Linux Bash Scripts/Scripts/Notion/visible.txt modified:   Linux Bash Scripts/Scripts/Rambox/check.txt modified:   Linux Bash Scripts/Scripts/Rambox/vis.txt modified:   Linux Bash Scripts/Scripts/Rambox/visible.txt modified:   Linux Bash Scripts/Scripts/Spotify/adjusted/check.txt modified:   Linux Bash Scripts/Scripts/Spotify/adjusted/vis.txt modified:   Linux Bash Scripts/Scripts/Spotify/adjusted/visible.txt modified:   Linux Bash Scripts/Scripts/Spotify_Volume/down.py modified:   Linux Bash Scripts/Scripts/Spotify_Volume/text.txt modified:   Linux Bash Scripts/Scripts/Spotify_Volume/text2.txt modified:   Linux Bash Scripts/Scripts/Spotify_Volume/up.py modified:   Linux Bash Scripts/Scripts/Temperature/Temperature.txt modified:   Linux Bash Scripts/Scripts/Temperature/fin.txt modified:   Linux Bash Scripts/Scripts/Terminate/kill.sh modified:   Linux Bash Scripts/Scripts/TestInternetSpeed/output.txt modified:   Linux Bash Scripts/Scripts/TestInternetSpeed/text1.txt modified:   Linux Bash Scripts/Scripts/Upgrade/Upgrade.py modified:   Linux Bash Scripts/Scripts/Vivaldi/Volume/text.txt modified:   Linux Bash Scripts/Scripts/Vivaldi/Volume/text2.txt modified:   Linux Bash Scripts/Scripts/Vivaldi/Volume/text3.txt modified:   Linux Bash Scripts/Scripts/Vivaldi/Volume/vol.py

+++ Linux Bash Scripts/Scripts/MicVolume/MicVolume.sh
@@ -1,24 +1,35 @@
 #!/bin/bash
 #include <math.h>
 
-in=$(zenity --entry --entry-text="65" --text="Enter % MicVolume")
 
-if [ $in != "" ]
-then
+
 ch=$(pacmd stat | grep "Default source name")
 
 source=${ch: 21}
 
 
+if [ "$1" = "" ];
+then
+
+in=$(zenity --entry --entry-text="65" --text="Enter % MicVolume")
+
+if [ "$in" = "" ];
+then
+
+pkill MicVolume.sh
+exit
+
+fi
+
+else
+
+in=95
+
+fi
+
 num=$(( $in*40000 / 61 ))
-# echo $num
+
 while sleep 0.1;
 do
 pacmd set-source-volume $source $num;
 done
-
-else
-
-pkill MicVolume.sh
-
-fi
